Move tests to aries-vcx-core
