+++ aries_vcx_core/src/utils/author_agreement.rs
@@ -51,3 +51,63 @@ pub fn get_global_txn_author_agreement() -> VcxCoreResult<Option<TxnAuthorAgreem
         Err(_) => Ok(None),
     }
 }
+
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used)]
+mod unit_tests {
+    use super::*;
+
+    const TEXT: &str = "indy agreement";
+    const VERSION: &str = "1.0.0";
+    const ACCEPTANCE_MECHANISM: &str = "acceptance mechanism label 1";
+    const TIME_OF_ACCEPTANCE: u64 = 123456789;
+
+    #[test]
+    fn set_txn_author_agreement_works() {
+        settings::reset_config_values().unwrap();
+        assert!(settings::get_config_value(settings::CONFIG_TXN_AUTHOR_AGREEMENT).is_err());
+
+        set_global_txn_author_agreement(
+            Some(TEXT.to_string()),
+            Some(VERSION.to_string()),
+            None,
+            ACCEPTANCE_MECHANISM.to_string(),
+            TIME_OF_ACCEPTANCE,
+        )
+            .unwrap();
+
+        assert!(settings::get_config_value(settings::CONFIG_TXN_AUTHOR_AGREEMENT).is_ok());
+    }
+
+    #[test]
+    fn get_txn_author_agreement_works() {
+        settings::reset_config_values().unwrap();
+        set_global_txn_author_agreement(
+            Some(TEXT.to_string()),
+            Some(VERSION.to_string()),
+            None,
+            ACCEPTANCE_MECHANISM.to_string(),
+            TIME_OF_ACCEPTANCE,
+        )
+            .unwrap();
+
+        let meta = get_global_txn_author_agreement().unwrap().unwrap();
+
+        let expected_meta = TxnAuthorAgreementAcceptanceData {
+            text: Some(TEXT.to_string()),
+            version: Some(VERSION.to_string()),
+            taa_digest: None,
+            acceptance_mechanism_type: ACCEPTANCE_MECHANISM.to_string(),
+            time_of_acceptance: TIME_OF_ACCEPTANCE,
+        };
+
+        assert_eq!(expected_meta, meta);
+    }
+
+    #[test]
+    fn get_txn_author_agreement_works_for_not_set() {
+        settings::reset_config_values().unwrap();
+        assert!(get_global_txn_author_agreement().unwrap().is_none());
+    }
+}
